apps/optimizer/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pulp>=2.9.0
numpy>=2.0.0
pandas>=2.2.0
python-docx>=1.1.0
supabase>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
matplotlib>=3.9.0
anthropic>=0.42.0
openpyxl>=3.1.0
python-pptx>=1.0.0
python-multipart>=0.0.20
pytest>=8.0.0
pytest-asyncio>=0.24.0
